requirements.txt
google-generativeai==0.3.0
flask==3.0.0
